requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
bootstrap4==0.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
fanstatic==1.1
gunicorn==20.0.4
idna==2.8
isort==4.3.21
js.jquery==3.3.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
shutilwhich==1.1.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
WebOb==1.8.6
wrapt==1.11.2